modified configure-system.sh

+++ Lubuntu/configure-system.sh
@@ -19,10 +19,10 @@ sudo apt-get build-dep --yes git
 (cd git && git clone https://github.com/git/git.git && cd git && make configure && ./configure && make)
 
 # clone&build egit/jgit
-(cd git && git clone https://git.eclipse.org/r/p/jgit/jgit && cd jgit && mvn clean install -DskipTests && mvn -f org.eclipse.jgit.packaging/pom.xml clean install)
-(cd git && git clone https://git.eclipse.org/r/p/egit/egit && cd egit && mvn -P skip-ui-tests clean install -DskipTests)
-(cd git && git clone https://git.eclipse.org/r/p/egit/egit-github && cd egit-github && mvn clean install)
-(cd git && git clone https://git.eclipse.org/r/p/egit/egit-pde && cd egit-pde && mvn clean install)
+(cd git && git clone https://git.eclipse.org/r/p/jgit/jgit && cd jgit && git config remote.origin.push HEAD:refs/for/master && mvn clean install -DskipTests && mvn -f org.eclipse.jgit.packaging/pom.xml clean install)
+(cd git && git clone https://git.eclipse.org/r/p/egit/egit && cd egit && git config remote.origin.push HEAD:refs/for/master && mvn -P skip-ui-tests clean install -DskipTests)
+(cd git && git clone https://git.eclipse.org/r/p/egit/egit-github && cd egit-github && git config remote.origin.push HEAD:refs/for/master && mvn clean install)
+(cd git && git clone https://git.eclipse.org/r/p/egit/egit-pde && cd egit-pde && git config remote.origin.push HEAD:refs/for/master && mvn clean install)
 
 # install eclipse juno
 (
@@ -116,17 +116,30 @@ chmod +x ~/no_proxy.sh
 # configured for chris
 git config --global user.name "Christian Halstrick"
 git config --global user.email "christian.halstrick@gmail.com"
-echo "machine git.eclipse.org" >> ~/.netrc
-echo "login chalstrick" >> ~/.netrc
-read -s -p "Enter password for chalstrick at git.eclipse.org: " epass
-echo "password $epass" >> ~/.netrc
-echo >> ~/.netrc
-echo "machine github.com" >> ~/.netrc
-echo "login chalstrick" >> ~/.netrc
-read -s -p "Enter password for chalstrick at github.com: " epass
-echo "password $epass" >> ~/.netrc
+if [ -f ~/.netrc ] ;then
+	echo "Don't write ~/.netrc because it already exists"
+else
+	read -s -p "Enter password for chalstrick at git.eclipse.org: " epass
+	echo -e "machine git.eclipse.org\nlogin chalstrick\npassword $epass" >> ~/.netrc
+	read -s -p "Enter password for chalstrick at github.com: " epass
+	echo -e "machine github.com\nlogin chalstrick\npassword $epass" >> ~/.netrc
+	chmod 600 ~/.netrc
+fi
 (cd git && git clone http://github.com/chalstrick/chrisFiles)
-cp git/chrisFiles/git/.gitconfig ~
-chmod 600 ~/.netrc
+if [ -f ~/.netrc ] ;then
+	echo "Don't write ~/.gitconfig because it already exists"
+else
+	cp git/chrisFiles/git/.gitconfig ~
+fi
+if [ -f ~/.vimrc ] ;then
+	echo "Don't write ~/.vimrc because it already exists"
+else
+	cp git/chrisFiles/vim/_vimrc ~/.vimrc
+fi
+
+(cd git/jgit && git config remote.origin.pushurl https://chalstrick@git.eclipse.org/r/p/jgit/jgit)
+(cd git/egit && git config remote.origin.pushurl https://chalstrick@git.eclipse.org/r/p/egit/egit)
+(cd git/egit-pde && git config remote.origin.pushurl https://chalstrick@git.eclipse.org/r/p/egit/egit-pde)
+(cd git/egit-github && git config remote.origin.pushurl https://chalstrick@git.eclipse.org/r/p/egit/egit-github)
 
 wait
